+++ st2actions/st2actions/runners/windows_command_runner.py
@@ -51,6 +51,8 @@ class WindowsCommandRunner(BaseWindowsRunner):
         self._timeout = timeout
 
     def pre_run(self):
+        super(WindowsCommandRunner, self).pre_run()
+
         # TODO :This is awful, but the way "runner_parameters" and other variables get
         # assigned on the runner instance is even worse. Those arguments should
         # be passed to the constructor.