+++ st2api/tests/base.py
@@ -93,6 +93,7 @@ class APIControllerWithRBACTestCase(FunctionalTest, CleanDbTestCase):
         super(APIControllerWithRBACTestCase, self).setUp()
 
         self.users = {}
+        self.roles = {}
 
         # Run RBAC migrations
         run_all_rbac_migrations()
@@ -108,6 +109,11 @@ class APIControllerWithRBACTestCase(FunctionalTest, CleanDbTestCase):
                 user=user_db.name,
                 role=role_name)
             UserRoleAssignment.add_or_update(role_assignment_db)
+
+        # Insert a user with no permissions and role assignments
+        user_1_db = UserDB(name='no_permissions')
+        user_1_db = User.add_or_update(user_1_db)
+        self.users['no_permissions'] = user_1_db
 
     def tearDown(self):
         super(APIControllerWithRBACTestCase, self).tearDown()