+++ st2common/st2common/constants/runners.py
@@ -18,8 +18,9 @@ from oslo_config import cfg
 __all__ = [
     'LOCAL_RUNNER_DEFAULT_ACTION_TIMEOUT',
 
-    'FABRIC_RUNNER_DEFAULT_ACTION_TIMEOUT',
-    'FABRIC_RUNNER_DEFAULT_REMOTE_DIR',
+    'REMOTE_RUNNER_DEFAULT_ACTION_TIMEOUT',
+    'REMOTE_RUNNER_DEFAULT_REMOTE_DIR',
+    'REMOTE_RUNNER_PRIVATE_KEY_HEADER',
 
     'PYTHON_RUNNER_DEFAULT_ACTION_TIMEOUT',
 
@@ -32,13 +33,15 @@ __all__ = [
 # Local runner
 LOCAL_RUNNER_DEFAULT_ACTION_TIMEOUT = 60
 
-# Remote (fabric runner)
-FABRIC_RUNNER_DEFAULT_ACTION_TIMEOUT = 60
+# Remote runner
+REMOTE_RUNNER_DEFAULT_ACTION_TIMEOUT = 60
 
 try:
-    FABRIC_RUNNER_DEFAULT_REMOTE_DIR = cfg.CONF.ssh_runner.remote_dir
+    REMOTE_RUNNER_DEFAULT_REMOTE_DIR = cfg.CONF.ssh_runner.remote_dir
 except:
-    FABRIC_RUNNER_DEFAULT_REMOTE_DIR = '/tmp'
+    REMOTE_RUNNER_DEFAULT_REMOTE_DIR = '/tmp'
+
+REMOTE_RUNNER_PRIVATE_KEY_HEADER = 'PRIVATE KEY-----'.lower()
 
 # Python runner
 # Default timeout (in seconds) for actions executed by Python runner