+++ CHANGELOG.rst
@@ -13,7 +13,15 @@ in development
   Note: This is an alternative to the existing ``--debug`` CLI flag which comes handy when running
   API services under gunicorn. (improvement)
 * Fix for `data` is dropped if `message` is not present in notification. (bug-fix)
-
+* Remove now deprecated Fabric based remote runner and corresponding
+  ``ssh_runner.use_paramiko_ssh_runner`` config option. (cleanup)
+* Fix support for password protected private key files in the remote runner. (bug-fix)
+* Allow user to provide a path to the private SSH key file for the remote runner ``private_key``
+  parameter. Previously only raw key material was supported. (improvement)
+* Add new API endpoint and corresponding CLI commands (``st2 runner disable <name>``,
+  ``st2 runner enable <name>``) which allows administrator to disable (and re-enable) a runner.
+  (new feature)
+* Add RBAC support for runner types API endpoints. (improvement)
 
 1.4.0 - April 18, 2016
 ----------------------